+++ UpnormalVmenu/app/src/androidTest/java/com/voidream/vmenu/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.voidream.upnormalvmenu;
+package com.voidream.vmenu;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
