+++ UpnormalVmenu/app/src/main/java/com/voidream/vmenu/MenuCustomAdapter.java
@@ -1,15 +1,12 @@
-package com.voidream.upnormalvmenu;
+package com.voidream.vmenu;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
-import android.graphics.Typeface;
-import android.hardware.camera2.params.StreamConfigurationMap;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
-import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
