+++ UpnormalVmenu/app/src/main/java/com/voidream/vmenu/SuksesMemesan.java
@@ -1,11 +1,10 @@
-package com.voidream.upnormalvmenu;
+package com.voidream.vmenu;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.ImageButton;
 
 public class SuksesMemesan extends AppCompatActivity {
