+++ UpnormalVmenu/app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:background="#f5f5f5"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.voidream.upnormalvmenu.MainActivity">
+    tools:context="com.voidream.vmenu.MainActivity">
 
     <TextView
         android:id="@+id/textOne"
