+++ UpnormalVmenu/app/src/main/res/layout/activity_pencarian.xml
@@ -7,7 +7,7 @@
     android:background="@android:color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.voidream.upnormalvmenu.Pencarian">
+    tools:context="com.voidream.vmenu.Pencarian">
 
     <TextView
         android:id="@+id/textOne"
