+++ UpnormalVmenu/app/src/main/res/layout/activity_pesan_sekarang.xml
@@ -8,7 +8,7 @@
     android:background="@android:color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.voidream.upnormalvmenu.PesanSekarang">
+    tools:context="com.voidream.vmenu.PesanSekarang">
 
     <RelativeLayout
         android:layout_width="match_parent"
