+++ UpnormalVmenu/app/src/main/res/layout/activity_popup_jumlah_pesanan.xml
@@ -4,7 +4,7 @@
     android:id="@+id/activity_popup_jumlah_pesanan"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.voidream.upnormalvmenu.PopupJumlahPesanan"
+    tools:context="com.voidream.vmenu.PopupJumlahPesanan"
     android:background="@color/black54">
 
     <LinearLayout
