+++ UpnormalVmenu/app/src/main/res/layout/activity_sukses_memesan.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context="com.voidream.upnormalvmenu.SuksesMemesan">
+    tools:context="com.voidream.vmenu.SuksesMemesan">
 
     <ImageView
         android:layout_width="match_parent"
